text matching init
