+++ tex/README.md
@@ -1,0 +1,30 @@
+# metadata-extraction
+
+
+## Requirements
+
+### Python Packages
+
+```
+texsoup
+
+```
+
+### Jupyter Extensions
+
+```
+jupytext
+```
+
+With **jupytext** you will be able to reconstitute the .py file
+ (a lightscript formatted file) back into a jupyter notebook. 
+ 
+
+
+## Setup
+
+### Data
+
+Create a folder called `data` at the top level of the repo. 
+It will be gitignored. 
+Put your local copy of the data files here.
