+++ text/samples/test.txt
@@ -1,0 +1,1 @@
+{"id":"https://ror.org/00tjv0s33","name":"Keimyung University","email_address":"","ip_addresses":[],"established":1899,"types":["Education"],"relationships":[{"label":"Keimyung University Dongsan Medical Center","type":"Related","id":"https://ror.org/035r7hb75"}],"addresses":[{"lat":35.857674,"lng":128.485571,"state":null,"state_code":null,"city":"Daegu","geonames_city":{"id":1835329,"city":"Daegu","geonames_admin1":{"name":"Daegu","id":1835327,"ascii_name":"Daegu","code":"KR.15"},"geonames_admin2":{"name":null,"id":null,"ascii_name":null,"code":null},"license":{"attribution":"Data from geonames.org under a CC-BY 3.0 license","license":"http://creativecommons.org/licenses/by/3.0/"},"nuts_level1":{"name":null,"code":null},"nuts_level2":{"name":null,"code":null},"nuts_level3":{"name":null,"code":null}},"postcode":null,"primary":false,"line":null,"country_geonames_id":1835841}],"links":["http://www.kmu.ac.kr/main.jsp"],"aliases":["Kei-dae"],"acronyms":["KMU"],"status":"active","wikipedia_url":"http://en.wikipedia.org/wiki/Keimyung_University","labels":[{"label":"계명대학교","iso639":"ko"}],"country":{"country_name":"South Korea","country_code":"KR"},"external_ids":{"ISNI":{"preferred":null,"all":["0000 0001 0669 3109"]},"FundRef":{"preferred":"501100002509","all":["501100002509","501100006017"]},"OrgRef":{"preferred":null,"all":["564378"]},"Wikidata":{"preferred":null,"all":["Q483402"]},"GRID":{"preferred":"grid.412091.f","all":"grid.412091.f"}}}
